refactor(spotify): :recycle: Split spotify data service into mulitple smaller
