--- libs/web/spotify/homepage/view/src/lib/artist/artist-albums/artist-albums.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ArtistAlbumsComponent } from './artist-albums.component';
-
-describe('ArtistAlbumsComponent', () => {
-  let component: ArtistAlbumsComponent;
-  let fixture: ComponentFixture<ArtistAlbumsComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ArtistAlbumsComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ArtistAlbumsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
