+++ libs/web/spotify/homepage/view/src/lib/search/search/search.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 import { CardListViewMode } from '@artur-ba/web/spotify/shared/view';
-import { SpotifyDataService } from '@artur-ba/web/spotify/shared/service';
+import { SpotifySearchDataService } from '@artur-ba/web/spotify/shared/service';
 
 @Component({
   selector: 'artur-ba-search',
@@ -18,16 +18,15 @@ export class SearchComponent {
   readonly albumsWrapperTitle = $localize`:search.albumsWrapper:Albums`;
   readonly playlistWrapperTitle = $localize`:search.playlistsWrapper:Playlists`;
 
-  constructor(protected readonly spotifyData: SpotifyDataService) {}
+  constructor(protected readonly spotifySearchData: SpotifySearchDataService) {}
 
   async search(searchQuery: string): Promise<void> {
     this.queryParams.q = encodeURI(searchQuery);
-    this.albumSearchResult = await this.spotifyData.getSearchAlbumResult(
+    this.albumSearchResult = await this.spotifySearchData.getSearchAlbumResult(
       searchQuery,
     );
-    this.playlistSearchResult = await this.spotifyData.getSearchPlaylistResult(
-      searchQuery,
-    );
+    this.playlistSearchResult =
+      await this.spotifySearchData.getSearchPlaylistResult(searchQuery);
   }
 
   getAlbumsWrapperTitle(): string {
