+++ libs/web/spotify/shared/service/src/index.ts
@@ -1,5 +1,8 @@
 export * from './lib/web-spotify-shared-service.module';
-
-export * from './lib/spotify-data.service';
 export * from './lib/player-control.service';
 export * from './lib/player.service';
+export * from './lib/spotify-data/spotify-data.service';
+export { SpotifyAlbumDataService } from './lib/spotify-data/spotify-album-data.service';
+export { SpotifyArtistDataService } from './lib/spotify-data/spotify-artist-data.service';
+export { SpotifyPlaylistDataService } from './lib/spotify-data/spotify-playlist-data.service';
+export { SpotifySearchDataService } from './lib/spotify-data/spotify-search-data.service';
