--- libs/web/spotify/shared/service/src/lib/player.service.spec.ts
@@ -1,17 +1,0 @@
-/// <reference types="spotify-web-playback-sdk" />
-import { TestBed } from '@angular/core/testing';
-
-import { PlayerService } from './player.service';
-
-describe('PlayerService', () => {
-  let service: PlayerService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PlayerService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
