+++ libs/web/spotify/shared/service/src/lib/player.service.ts
@@ -1,4 +1,3 @@
-/// <reference types="spotify-web-playback-sdk" />
 import { Injectable, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { Title } from '@angular/platform-browser';
