--- libs/web/spotify/shared/service/src/lib/spotify-data.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SpotifyDataService } from './spotify-data.service';
-
-describe('SpotifyDataService', () => {
-  let service: SpotifyDataService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SpotifyDataService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
